// src/Mobile/ReportsApp/services/api/models/UpdateCheckListItemCommand.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Inspections API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    CheckValue,
    CheckValueFromJSON,
    CheckValueFromJSONTyped,
    CheckValueToJSON,
} from './';

/**
 * 
 * @export
 * @interface UpdateCheckListItemCommand
 */
export interface UpdateCheckListItemCommand {
    /**
     * 
     * @type {number}
     * @memberof UpdateCheckListItemCommand
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateCheckListItemCommand
     */
    checkListId?: number;
    /**
     * 
     * @type {string}
     * @memberof UpdateCheckListItemCommand
     */
    text?: string | null;
    /**
     * 
     * @type {CheckValue}
     * @memberof UpdateCheckListItemCommand
     */
    checked?: CheckValue;
    /**
     * 
     * @type {boolean}
     * @memberof UpdateCheckListItemCommand
     */
    editable?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof UpdateCheckListItemCommand
     */
    required?: boolean;
    /**
     * 
     * @type {string}
     * @memberof UpdateCheckListItemCommand
     */
    remarks?: string | null;
}

export function UpdateCheckListItemCommandFromJSON(json: any): UpdateCheckListItemCommand {
    return UpdateCheckListItemCommandFromJSONTyped(json, false);
}

export function UpdateCheckListItemCommandFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateCheckListItemCommand {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'checkListId': !exists(json, 'checkListId') ? undefined : json['checkListId'],
        'text': !exists(json, 'text') ? undefined : json['text'],
        'checked': !exists(json, 'checked') ? undefined : CheckValueFromJSON(json['checked']),
        'editable': !exists(json, 'editable') ? undefined : json['editable'],
        'required': !exists(json, 'required') ? undefined : json['required'],
        'remarks': !exists(json, 'remarks') ? undefined : json['remarks'],
    };
}

export function UpdateCheckListItemCommandToJSON(value?: UpdateCheckListItemCommand | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'checkListId': value.checkListId,
        'text': value.text,
        'checked': CheckValueToJSON(value.checked),
        'editable': value.editable,
        'required': value.required,
        'remarks': value.remarks,
    };
}